+++ src/Adapter/FilesystemAdapter.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Dot\Cache\Adapter;
+
+use Symfony\Component\Cache\Adapter;
+
+class FilesystemAdapter extends Adapter\FilesystemAdapter
+{
+}